refactor: split organization manager
