+++ src/controllers/main.ts
@@ -2,10 +2,10 @@ import { NextFunction, Request, Response } from 'express';
 import getNotificationsFromRequest from '../services/get-notifications-from-request';
 import { ZodError } from 'zod';
 import { updatePersonalInformations } from '../managers/user';
-import { getUserOrganizations } from '../managers/organization';
 import notificationMessages from '../notification-messages';
 import { getClientsOrderedByConnectionCount } from '../managers/oidc-client';
 import { getParamsForPostPersonalInformationsController } from './user/update-personal-informations';
+import { getUserOrganizations } from '../managers/organization/main';
 
 export const getHomeController = async (
   req: Request,
